בהצלחה

+++ Hever/Models/Users.cs
@@ -11,7 +11,7 @@ namespace Hever.Models
         public string UserName { get; set; }
         public int Password { get; set; }
         public bool IsAdmine { get; set; }
-        public int CardNumber { get; set; }
+        public long CardNumber { get; set; }
         public DateTime CardExpirationDate { get; set; }
         public int Cvs { get; set; }
         public ICollection<Restaurant> LikedRestaurants { get; set; }
